Added comm parsing and skeletal input

+++ LucidGloves/LucidGloves/include/Comm/CommunicationManager.h
@@ -4,7 +4,7 @@
 class ICommunicationManager {
 public:
 	virtual void Connect() = 0;
-	virtual void BeginListener(const std::function<void(const int*)>& callback) = 0;
+	virtual void BeginListener(const std::function<void(const float*)>& callback) = 0;
 	virtual bool IsConnected() = 0;
 	virtual void Disconnect() = 0;
 	virtual ~ICommunicationManager() {};
